# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Gerenciamento de Academias",
    "version": "1.0.0",
    "description": "Documentação da API de login e rotas protegidas para Aluno, Professor e Admin"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/login/professor": {
      "post": {
        "tags": ["Professor"],
        "summary": "Login de professor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "professor@email.com"
                  },
                  "password": { "type": "string", "example": "senha123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
...
    "/professor": {
      "get": {
        "tags": ["Professor"],
        "summary": "Área exclusiva do professor",
        "responses": {
          "200": { "description": "Acesso autorizado" },
          "401": { "description": "Não autorizado" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}